fix warn

+++ Siv3D/Main.cpp
@@ -13,18 +13,18 @@ void Main()
     {
         for (size_t i = 0; i < 8; ++i) {
             HSL hsl = ColorCast<HSL>(base);
-            hsl.l = s3d::Math::Lerp(0.0, 1.0, s3d::Math::InvLerp(0, 7, i));
+            hsl.l = s3d::Math::Lerp(0.0, 1.0, s3d::Math::InvLerp(0, 7, static_cast<double>(i)));
             Rect{ 100 * i, 000, 100, 100 }.draw(ColorCast<ColorF>(hsl));
         }
         for (size_t i = 0; i < 8; ++i) {
             LCH lch = ColorCast<LCH>(base);
-            lch.l = s3d::Math::Lerp(0.0, 100.0, s3d::Math::InvLerp(0, 7, i));
+            lch.l = s3d::Math::Lerp(0.0, 100.0, s3d::Math::InvLerp(0, 7, static_cast<double>(i)));
             lch = colorcpp::Utility::ClampChroma(lch);
             Rect{ 100 * i, 100, 100, 100 }.draw(ColorCast<ColorF>(lch));
         }
         for (size_t i = 0; i < 8; ++i) {
             OkLCH lch = ColorCast<OkLCH>(base);
-            lch.l = s3d::Math::Lerp(0.0, 1.0, s3d::Math::InvLerp(0, 7, i));
+            lch.l = s3d::Math::Lerp(0.0, 1.0, s3d::Math::InvLerp(0, 7, static_cast<double>(i)));
             lch = colorcpp::Utility::ClampChroma(lch);
             Rect{ 100 * i, 200, 100, 100 }.draw(ColorCast<ColorF>(lch));
         }
